# doc/api.yaml
openapi: 3.0.0
info:
  title: WASATEXT
  description: API for WASATEXT messaging platform
  version: 1.0.0
...
security:
  - WASATextAuth: []
servers:
  - url: http://0.0.0.0:3000
    description: Local Server
paths:
...
  /users/{userId}/info/username:
    put:
      operationId: setMyUserName
      summary: Set personalized username
      description: Let the user change his username into another.
        If the new username is available, set the new username. Otherwise, keep the last username saved and returns error.
      tags:
        - "User"
      parameters:
        - $ref: "#/components/parameters/userId"
      requestBody:
        description: New username proposed
        required: true
        content:
          application/json:
            schema:
              description: Propose's schema
              type: object
              properties:
                newusername:
                  $ref: "#/components/schemas/Username"

      responses:
        "201":
          description: Edit successful
          content:
            application/json:
              schema:
                description: Response's schema
                type: object
                properties:
                  userId:
                    $ref: "#/components/schemas/UserId"
                  username:
                    $ref: "#/components/schemas/Username"
              example:
                userId: "19935"
                username: simone

        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          description: username already in use
          content:
            application/json:
              schema:
                description: Response's schema
                type: object
                properties:
                  userId:
                    $ref: "#/components/schemas/UserId"
                  username:
                    $ref: "#/components/schemas/Username"
              example:
                userId: "19935"
                username: simone
        "500":
          $ref: "#/components/responses/InternalError"
...
  /users/{userId}/groups:
    parameters:
      - $ref: "#/components/parameters/userId"
    get:
      summary: View all groups
      description: Returns the list of groups
      operationId: viewGroups
      tags:
        - "Group"
      responses:
        "200":
          description: OK. Return The list
          content:
            application/json:
              schema:
                description: List of groups
                properties:
                  groups:
                    description: Groups schema
                    type: array
                    minItems: 0
                    maxItems: 2048
                    items:
                      allOf:
                        - $ref: "#/components/schemas/Group"
                        - properties:
                            lastMessage:
                              $ref: "#/components/schemas/Message"
                example:
                  groups:
                    - groupId: "288344"
                      name: "capodanno2026"
                      upload:
                        photoId: "88332"
                        url: "https://www.stoccolmaaroma.it/wp-content/uploads/2013/07/logo-sapienza.jpg"
                        mime: "image/jpg"
                        width: 1024
                        height: 720
                      participants:
                        - userId: "10012"
                          username: "Simone"
                        - userId: "77772"
                          username: "Alessio"
    post:
      summary: Create new group
      description: Create a new group conversation.
      operationId: createGroup
      tags:
        - "Group"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Request schema
              properties:
                name:
                  description: Group name
                  type: string
                  minLength: 0
                  maxLength: 255
                  pattern: "^[a-zA-Z0-9_ -]+$"
                  example: capodanno2026
                participants:
                  description: list of participants of a group
                  type: array
                  minItems: 2
                  maxItems: 256
                  items:
                    $ref: "#/components/schemas/Username"
                  example:
                      - Simone
                      - Luca
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
              example:
                groupId: "288344"
                name: "capodanno2026"

                participants:
                  - userId: "10012"
                    username: "Simone"
                  - userId: "77772"
                    username: "Alessio"

        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /users/{userId}/groups/{groupId}/info/photo:
    put:
      summary: Set Group Photo
      description: The user can set a personalized group Photo.
      operationId: setGroupPhoto
      tags:
        - "Group"
      parameters:
        - $ref: "#/components/parameters/userId"
        - $ref: "#/components/parameters/GroupId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Upload photo's schema
              properties:
                photo:
                  $ref: "#/components/schemas/File"
              required:
                - photo
      responses:
        "201":
          description: Image updated.
          content:
            application/json:
              schema:
                description: Updated group Photo information
                properties:
                  upload:
                    $ref: "#/components/schemas/Upload"
                  time:
                    description: Timestamp
                    type: string
                    format: date-time
                    example: 2025-10-31T15:42:00Z
                    minLength: 20
                    maxLength: 20
                    pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}Z$"
              example:
                upload:
                  url: "https://example.com/uploads/groups/123/group_photo.png"
                  mime: "image/png"
                  width: 512
                  height: 512
                time: "2025-10-31T15:42:00Z"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /users/{userId}/info/photo:
    put:
      summary: Change Avatar
      description: The user can edit their profile Photo
      tags:
        - "User"
      operationId: setMyPhoto
      parameters:
        - $ref: "#/components/parameters/userId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Upload photo's schema
              properties:
                photo:
                  $ref: "#/components/schemas/File"
              required:
                - photo
      responses:
        "201":
          description: Image updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Upload"
              example:
                url: "https://example.com/avatar.jpg"
                mime: "image/png"
                width: 512
                height: 512
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /users/{userId}/groups/{groupId}:
    get:
      summary: Show the conversation
      description: Returns the details of a single conversation identified by conversationId from the user's conversations.
      operationId: getGroup
      tags:
        - "Conversations"
      parameters:
        - $ref: "#/components/parameters/userId"
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: Returns the chat as a list of messages.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
              example:
                groupId: "288344"
                name: "capodanno2026"
                upload:
                  url: "https://www.stoccolmaaroma.it/wp-content/uploads/2013/07/logo-sapienza.jpg"
                  mime: "image/jpg"
                  width: 1024
                  height: 720
                participants:
                  - userId: "10012"
                    username: "Simone"
                  - userId: "77772"
                    username: "Alessio"


        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  schemas:
...
    UserId:
      description: Operation User's ID
      type: string
      pattern: "^[0-9]+$"
      minLength: 1
      maxLength: 9999
      example: "11300"
      readOnly: true
    Username:
      description: User's username
      type: string
      minLength: 1
      maxLength: 255
      pattern: "^[a-zA-Z0-9_ -]+$"
      example: Simone
    User:
      description: User schema
      type: object
      properties:
        userId:
          $ref: "#/components/schemas/UserId"
        username:
          $ref: "#/components/schemas/Username"
      required:
        - userId
        - username
    Group:
      type: object
      description: group schema
      properties:
        groupId:
          $ref: "#/components/schemas/GroupId"
        name:
          description: Group name
          type: string
          minLength: 0
          maxLength: 255
          pattern: "^[a-zA-Z0-9_ -]+$"
          example: capodanno2026
        upload:
          $ref: "#/components/schemas/Upload"
        participants:
          description: list of participants of a group
          type: array
          minItems: 2
          maxItems: 256
          items:
            $ref: "#/components/schemas/User"
          example:
            - userId: "11300"
              username: Simone
            - userId: "11301"
              username: Luca
...
    MessageId:
      description: Unique ID assigned to each message resource in the system.
      pattern: "^[0-9]+$"
      minLength: 1
      maxLength: 9999
      example: "2635444"
      readOnly: true
      type: string
    GroupId:
      description: Unique ID assigned to each group resource in the system.
      type: string
      pattern: "^[0-9]+$"
      minLength: 1
      maxLength: 9999
      example: "2773354"
      readOnly: true
...
    File:
      description: Binary image file
      type: string
      format: binary
      pattern: "^[A-Za-z0-9+/=]+$"
      minLength: 1
      maxLength: 9999999999999999999
    Upload:
      description: upload's schema for responses or view
      type: object
      properties:
        url:
          type: string
          minLength: 1
          maxLength: 255
          pattern: "^[a-zA-Z]+://.+$"
          format: url
          description: Photo's URL to access the image
          example: https://www.stoccolmaaroma.it/wp-content/uploads/2013/07/logo-sapienza.jpg
        mime:
          description: mime represents the type of the Photo(image/png...)
          type: string
          minLength: 0
          maxLength: 255
          pattern: '^[a-zA-Z]+/[a-zA-Z.]+$'
          example: image/jpg
        width:
          description: width of the image
          type: integer
          example: 1024
        height:
          description: height of the image
          type: integer
          example: 720
    Message:
      type: object
      description: message schema
      properties:
        messageId:
          $ref: "#/components/schemas/MessageId"
        body:
          type: object
          description: messageContent
          properties:
            text:
              $ref: "#/components/schemas/Text"
            upload:
              $ref: "#/components/schemas/Upload"
        time:
          description: time
          type: string
          format: date-time
          example: "2025-10-31T15:42:00Z"
        sender:
          $ref: "#/components/schemas/User"
        read:
          description: show if the message has been read
          type: string
          enum: [ received, read ]
          example: received
        isForwarded:
          description: show if the message has been forwarded
          type: boolean
          example: false
        replyTo:
          $ref: "#/components/schemas/MessageId"
    Text:
      description: Text of the message
      type: string
      minLength: 0
      maxLength: 2048
      example: Hello, World!
      pattern: "^[a-zA-Z0-9 .,!?_-]+$"
...
    Error:
      description: Info of an error
      type: object
      properties:
        code:
          description: code of the error
          type: integer
          example: 401
        message:
          description: text of the error
          type: string
          minLength: 0
          maxLength: 2048
          example: Unauthorized Action, request invalid or missing.
      required:
        - code
        - message
  parameters:
    userId:
      name: userId
      in: path
      required: true
      description: User Identifier
      schema:
        $ref: "#/components/schemas/UserId"
...
    GroupId:
      name: groupId
      in: path
      required: true
      description: Group Identifier
      schema:
        $ref: "#/components/schemas/GroupId"
  securitySchemes:
    WASATextAuth:
      type: http
      description: Protect operation session token.
      scheme: bearer
      bearerFormat: SessionToken
  responses:
    UnauthorizedError:
      description: Unauthorized Action, request invalid or missing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 401
            message: "Unauthorized Action, request invalid or missing."
    InternalError:
      description: An error occurred, please retry.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 500
            message: "An error occurred, please retry."
    NotFoundError:
      description: Resource not found. Check the parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 404
            message: "Resource not found. Check the parameters."
    BadRequest:
      description: Bad Request. Invalid parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 400
            message: "Bad Request. Invalid parameters."
...